move PlayerState
