+++ packages/gamemode/src/entities/player/index.ts
@@ -1,3 +1,4 @@
 export { PlayerEntity, playerEntity } from "./entity";
 export { PlayerGameActions } from "./sagas";
 export * as PlayerEntitySelectors from "./selectors";
+export { PlayerState } from "./state";
