+++ packages/gamemode/src/entities/player/sagas/fillBoard.ts
@@ -1,6 +1,6 @@
 import { takeEvery, put } from "redux-saga/effects";
 import { getContext, select } from "typed-redux-saga";
-import { PlayerState } from "../../../player/store";
+import { PlayerState } from "../state";
 import { PlayerPieceLocation } from "@creature-chess/models";
 import { BoardSelectors } from "@creature-chess/board";
 import { getMostExpensiveBenchPiece, getPlayerBelowPieceLimit, isPlayerAlive } from "../../../player/playerSelectors";
