+++ packages/gamemode/src/index.ts
@@ -8,12 +8,11 @@ export {
 	PlayerVariables
 } from "./player";
 
-export { PlayerEntity, playerEntity, PlayerGameActions, PlayerEntitySelectors } from "./entities/player";
+export { PlayerEntity, playerEntity, PlayerGameActions, PlayerEntitySelectors, PlayerState } from "./entities/player";
 
 export { getPiece, getAllPieces } from "./player/pieceSelectors";
 export { getPlayerLevel, getPlayerMoney, getPlayerXp, isPlayerAlive } from "./player/playerSelectors";
 export { PlayerInfoCommands, PlayerInfoState, playerInfoReducer, HasPlayerInfo, PlayerMatchRewards } from "./player/playerInfo";
-export { PlayerState } from "./player/store";
 
 export { RoundInfoState, roundInfoReducer, RoundInfoCommands } from "./game/roundInfo";
 export * as GameEvents from "./game/events";
